Move modal's childrens to Modal folder
