+++ components/FoldersList/index.tsx
@@ -1,12 +1,12 @@
 import { ReactElement, useState } from "react";
 import { FolderData } from "@/lib/api";
-import FolderAddForm from "@/components/FolderAddForm";
-import SocialShareBox from "@/components/SocialShareBox";
-import FolderEditForm from "@/components/FolderEditForm";
-import FolderDeleteForm from "@/components/FolderDeleteForm";
 import Modal from "@/components/Modal";
 import { FaPencilAlt, FaRegShareSquare, FaRegTrashAlt } from "react-icons/fa";
 import styles from "@/components/FoldersList/FoldersList.module.css";
+import FolderAddForm from "../Modal/childrens/FolderAddForm";
+import SocialShareBox from "../Modal/childrens/SocialShareBox";
+import FolderEditForm from "../Modal/childrens/FolderEditForm";
+import FolderDeleteForm from "../Modal/childrens/FolderDeleteForm";
 
 interface FoldersListProps {
   handleClick: (folderId: number | null) => void;
