+++ components/Modal/childrens/FolderAddForm/index.tsx
@@ -1,4 +1,4 @@
-import styles from "@/components/Modal/ModalChildren.module.css";
+import styles from "@/components/Modal/childrens/ModalChildren.module.css";
 
 function FolderAddForm() {
   return (
