+++ components/Modal/childrens/SocialShareBox/index.tsx
@@ -1,5 +1,5 @@
 import Image from "next/image";
-import styles from "@/components/Modal/ModalChildren.module.css";
+import styles from "@/components/Modal/childrens/ModalChildren.module.css";
 
 interface SocialShareBoxProps {
   title: string;
